♻️ 71SCSW task: extract integration queue routing
