+++ .agentplane/tasks/202607221852-71SCSW/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221852-71SCSW/extend-supervised-execution-to-branch-pr",
   "created_at": "2026-07-31T11:14:18.745Z",
-  "diffstat_sha256": "sha256:bb0d09d8dd284d96bf01a67d76f289f9c79a3b9ce73f13b606a13ff5742b62df",
+  "diffstat_sha256": "sha256:92e30dca65a893a98f699bad796cf49c8de00381035ea49c560b33cb795e008b",
   "last_verified_at": "2026-07-31T12:09:21.598Z",
   "last_verified_diffstat_sha256": "sha256:ebe794e10f692485c248a5ee84913e6aa0326ae3eba32ee88c4099117d87d4de",
   "pr_number": 4702,
